requirements.txt
google-generativeai
python-dotenv
scikit-learn
pypdf2
pyTelegramBotAPI